pyproject.toml
[project]
name = "tissueagent"
version = "0.1.0"
description = "autonomous AI agent for spatial transcriptomics research"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain==0.3.25",
    "langchain-core==0.3.64",
    "langchain-ollama==0.3.0",
    "langchain-openai==0.3.10",
    "langchain-anthropic>=0.3.15",
    "langchain-community==0.3.24",
    "langgraph==0.3.20",
    "matplotlib==3.9.4",
    "numpy>=2.0.2",
    "pandas==2.2.3",
    "pydantic==2.10.6",
    "scanpy==1.10.3",
    "scipy==1.13.1",
    "seaborn==0.13.2",
    "spatialdata-io==0.1.5",
    "squidpy>=1.6.2",
    "streamlit==1.43.2",
    "streamlit-nested-layout==0.1.4",
    "streamlit-file-browser==3.2.22",
    "streamlit-modal==0.1.2",
    "tiledbsoma==1.17.1",
    "google-search-results==2.4.2",
    "pymed==0.8.9",
    "watchdog>=6.0.0",
    "anthropic==0.52.0",
    "openai>=1.73.0",
    "nbformat>=5.10.4",
    "langchain-chroma==0.2.4",
    "langchain-experimental>=0.3.0",
    "colorama>=0.4.6",
    "dask==2024.8.0",
    "xarray-datatree==0.0.15",
    "cellxgene-census>=1.17.0",
    "cell2location>=0.1.4",
    "mcp[cli]>=1.12.4",
    "sentence-transformers>=2.2.0",
    "scikit-learn>=1.3.0",
    "nbconvert>=7.16.6",
    "anndata>=0.11.4",
    "liana>=1.6.1",
    "decoupler>=2.1.1",
    "notebook>=7.4.7",
    "marimo>=0.17.2",
    "ggplot>=0.11.5",
    "plotnine>=0.15.0",
    "memorisdk==2.3.0",
    "ruff>=0.15.6",
    "pre-commit>=4.5.1",
]

[tool.ruff]
line-length = 120
force-exclude = true

[tool.ruff.lint]
select = ["E", "F", "D"]
ignore = ["E731"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"**/prompt.py" = ["E501"]
"**/__init__.py" = ["ALL"]
"notebooks/**" = ["ALL"]
"**/deprecated/**" = ["ALL"]